FlaskApp/requirements.txt
flask
requests
request
google-cloud-bigquery[pandas]==1.9.0
google-auth-oauthlib==0.2.0
ipython==7.2.0
matplotlib==3.0.2
pytz==2018.9
